Move all issue-related code to an issue lib
